+++ image/service/clamav/process.sh
@@ -4,6 +4,8 @@
 # https://github.com/osixia/docker-light-baseimage/blob/stable/image/tool/log-helper
 log-helper level eq trace && set -x
 
+chown -R clamav:clamav /var/run/clamav
+
 ln -sf ${CONTAINER_SERVICE_DIR}/clamav/assets/cronjobs /etc/cron.d/clamav
 chmod 600 ${CONTAINER_SERVICE_DIR}/clamav/assets/cronjobs