+++ image/service/opendkim/install.sh
@@ -3,5 +3,3 @@
 
 # opendkim
 mkdir -p /etc/opendkim
-
-echo "SOCKET=\"inet:12301@127.0.0.1\"" >> /etc/default/opendkim