+++ image/service/postfix/startup.sh
@@ -14,9 +14,12 @@ if [ ! -e "$FIRST_START_DONE" ]; then
   # set mailserver hostname
   sed -i "s|{{ HOSTNAME }}|${HOSTNAME}|g" ${CONTAINER_SERVICE_DIR}/postfix/assets/config/main.cf
 
+  # set relayhost
+  sed -i "s|{{ MMC_MAIL_RELAYHOST }}|${MMC_MAIL_RELAYHOST}|g" ${CONTAINER_SERVICE_DIR}/postfix/assets/config/main.cf
+
   # generate a certificate and key if files don't exists
-  # https://github.com/osixia/docker-light-baseimage/blob/stable/image/service-available/:cfssl/assets/tool/cfssl-helper
-  cfssl-helper ${MMC_MAIL_CFSSL_PREFIX} "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_CRT_FILENAME" "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_KEY_FILENAME" "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_CA_CRT_FILENAME"
+  # https://github.com/osixia/docker-light-baseimage/blob/stable/image/service-available/:ssl-tools/assets/tool/ssl-helper
+  ssl-helper ${MMC_MAIL_SSL_HELPER_PREFIX} "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_CRT_FILENAME" "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_KEY_FILENAME" "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_CA_CRT_FILENAME"
 
   MMC_MAIL_SSL_CA_CRT_PATH="${CONTAINER_SERVICE_DIR}/postfix/assets/certs/${MMC_MAIL_SSL_CA_CRT_FILENAME}"
   MMC_MAIL_SSL_CRT_PATH="${CONTAINER_SERVICE_DIR}/postfix/assets/certs/${MMC_MAIL_SSL_CRT_FILENAME}"
@@ -29,6 +32,9 @@ if [ ! -e "$FIRST_START_DONE" ]; then
 
   MMC_MAIL_SSL_DHPARAM_512_PATH="${CONTAINER_SERVICE_DIR}/postfix/assets/certs/${MMC_MAIL_SSL_DHPARAM_512}"
   MMC_MAIL_SSL_DHPARAM_1024_PATH="${CONTAINER_SERVICE_DIR}/postfix/assets/certs/${MMC_MAIL_SSL_DHPARAM_1024}"
+
+  [ -f ${MMC_MAIL_SSL_DHPARAM_512_PATH} ] || openssl dhparam -out ${MMC_MAIL_SSL_DHPARAM_512_PATH} 512
+  [ -f ${MMC_MAIL_SSL_DHPARAM_1024_PATH} ] || openssl dhparam -out ${MMC_MAIL_SSL_DHPARAM_1024_PATH} 1024
 
   sed -i "s|{{ MMC_MAIL_SSL_DHPARAM_512_PATH }}|${MMC_MAIL_SSL_DHPARAM_512_PATH}|g" ${CONTAINER_SERVICE_DIR}/postfix/assets/config/main.cf
   sed -i "s|{{ MMC_MAIL_SSL_DHPARAM_1024_PATH }}|${MMC_MAIL_SSL_DHPARAM_1024_PATH}|g" ${CONTAINER_SERVICE_DIR}/postfix/assets/config/main.cf