+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.8
+  - Fix opendkim config
+  - Upgrade baseimage: light-baseimage:0.2.5
+  - Add tls certificates for replication only
+  - Add opendmarc
+
 ## 0.1.7
   - Upgrade baseimage: light-baseimage:0.2.2